[Feat][psor_books] Add psor_books CLI
The psor_books command acts as a wrapper around the jupyter book build command. If the option --publish is passed, it will strip out the REMOVE-FROM-PUBLISH sections from config and toc and place those in book/.psorbooks/publish. Then, jupyter book build is called with --config and --toc options pointing to the cleaned files.